tests/wpt/tests/html/canvas/offscreen/text/2d.text.drawing.style.measure.textAlign.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.text.drawing.style.measure.textAlign</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.text.drawing.style.measure.textAlign</h1>
<p class="desc">Measurement should be related to textAlignment</p>


<script>
var t = async_test("Measurement should be related to textAlignment");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.textAlign = "right";
  metrics = ctx.measureText('hello');
  _assert(metrics.actualBoundingBoxLeft > metrics.actualBoundingBoxRight, "metrics.actualBoundingBoxLeft > metrics.actualBoundingBoxRight");

  ctx.textAlign = "left"
  metrics = ctx.measureText('hello');
  _assert(metrics.actualBoundingBoxLeft < metrics.actualBoundingBoxRight, "metrics.actualBoundingBoxLeft < metrics.actualBoundingBoxRight");
  t.done();

});
</script>
